+++ components/Layout.js
@@ -18,6 +18,7 @@ class Layout extends React.PureComponent {
       <div>
         <Head>
           <script src="ace/src-min/ace.js" />
+          <script src="ace/src-min/ext-language_tools.js" />
           <link rel="stylesheet" type="text/css" href="hljs/styles/github.css" />
           <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/github-markdown-css/2.9.0/github-markdown.min.css" />
           <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />