+++ pages/markdown.js
@@ -37,6 +37,7 @@ export default class extends React.Component {
           <Editor
             content={this.props.markdownExample}
             language={markdown}
+            listFiles={() => Promise.resolve(['foo', 'bar', 'baz', 'quix'])}
           />
         </div>
       </Layout>