+++ src/components/Editor.js
@@ -18,18 +18,21 @@ export default class EmilyEditor extends React.PureComponent {
     content: PropTypes.string,
     language: PropTypes.shape({
       name: PropTypes.string.isRequired,
-      toHtml: PropTypes.func.isRequired,
+      convert: PropTypes.func.isRequired,
       lineSafeInsert: PropTypes.func.isRequired,
       postProcess: PropTypes.func.isRequired,
       renderJsxStyle: PropTypes.func.isRequired,
       excludeNode: PropTypes.func.isRequired,
       previewClassName: PropTypes.string.isRequired,
     }).isRequired,
+    // eslint-disable-next-line react/no-unused-prop-types
+    listFiles: PropTypes.func,
     width: PropTypes.number,
     height: PropTypes.number,
   }
   static defaultProps = {
     content: '',
+    listFiles: () => Promise.resolve([]),
     width: null,
     height: null,
   }
@@ -46,7 +49,7 @@ export default class EmilyEditor extends React.PureComponent {
       },
     };
 
-    const html = this.generateHtml(props.content);
+    const { html, suggestions } = this.generateHtml(props.content);
     const raw = props.content;
     this.state = {
       raw,
@@ -58,6 +61,8 @@ export default class EmilyEditor extends React.PureComponent {
         preview: true,
         outline: true,
       },
+      // eslint-disable-next-line react/no-unused-state
+      suggestions,
       aceOptions: defaultAceOptions,
       autosaved: null,
       loc: raw.split('\n').length,
@@ -136,11 +141,13 @@ export default class EmilyEditor extends React.PureComponent {
     this.scrollPreviewToLine(firstVisibleLine);
   }
   updateStateValue = (value) => {
-    const html = this.generateHtml(value);
+    const { html, suggestions } = this.generateHtml(value);
     const raw = value;
     this.setState({
       raw,
       html,
+      // eslint-disable-next-line react/no-unused-state
+      suggestions,
       loc: raw.split('\n').length,
       outline: generateOutline(html, raw, this.props.language.excludeNode),
     });
@@ -191,12 +198,14 @@ export default class EmilyEditor extends React.PureComponent {
     }
   }
   generateHtml = (raw) => {
-    const { lineSafeInsert, toHtml } = this.props.language;
-    const htmlWithNinjas = toHtmlWithNinjas(raw, lineSafeInsert, toHtml);
-    return applyOnDom(htmlWithNinjas, (node) => {
+    const { lineSafeInsert, convert } = this.props.language;
+    const { html, suggestions } = toHtmlWithNinjas(raw, lineSafeInsert, convert);
+    const processPrevewHtml = htmlSrc => applyOnDom(htmlSrc, (node) => {
       node.querySelectorAll('a').forEach(n => n.setAttribute('target', '_blank'));
       this.props.language.postProcess(node);
     });
+
+    return { html: processPrevewHtml(html), suggestions };
   }
   handleCommand = (command) => {
     getCommands(this)[command].execute();