+++ src/components/editor/lineNinja.js
@@ -17,8 +17,10 @@ export const addNinjas = (src, insert) => src
   .join('\n');
 
 
-export const toHtmlWithNinjas = (src, insert, toHtml) =>
-  ninjasToHtml(toHtml(addNinjas(src, insert)));
+export const toHtmlWithNinjas = (src, insert, convert) => {
+  const converted = convert(addNinjas(src, insert), src);
+  return { ...converted, html: ninjasToHtml(converted.html) };
+};
 
 export const higlightSourceWithNinjas = (source, highlightFn) => {
   const matches = source.match(ninjaRegex);