+++ test/modes/asciidoc.test.js
@@ -1,7 +1,8 @@
+import _ from 'lodash';
 import asciidoc from '../../src/modes/asciidoc';
 
 describe('Asciidoc mode', () => {
-  const documentFromSource = source => new DOMParser().parseFromString(asciidoc.toHtml(source), 'text/html').body;
+  const documentFromSource = source => new DOMParser().parseFromString(asciidoc.convert(source).html, 'text/html').body;
   const source = '= foo\nfoo\n\n== bar\nbar\n\n=== *baz*\n_quix_';
   const ninja = '{=O=O=}';
   const testContents = (src) => {
@@ -32,6 +33,15 @@ describe('Asciidoc mode', () => {
       it('Ninjas match number of lines', () => {
         expect(withNinjas.match(new RegExp(ninja, 'g')).length).toBe(source.split('\n').filter(x => x !== '').length);
       });
+    });
+  });
+  describe('Autocomplete', () => {
+    const refs = _.flatMap(asciidoc.convert(source).suggestions, x => _.get(x, 'refs'));
+    it('Contains plain heading reference', () => {
+      expect(_.some(refs, { value: '_bar, bar', caption: 'bar', meta: 'reference' })).toBe(true);
+    });
+    it('Contains html heading reference with stripped caption', () => {
+      expect(_.some(refs, { value: '_strong_baz_strong, baz', caption: 'baz', meta: 'reference' })).toBe(true);
     });
   });
 });